Collect the registered queries in the compiler
