+++ lib/hologram/compiler.ex
@@ -303,6 +303,56 @@ defmodule Hologram.Compiler do
   end
 
   @doc """
+  Collects the registered queries of the given component modules and derives what a running
+  Hologram app holds for them: the registry entries keyed by query content id, the argument names
+  of every parameterized capture keyed by `{module, prop name}`, and the term behind each window
+  keyed by window id.
+
+  The component modules are every component of the build rather than the ones pages reach, because
+  what reads the result is the renderer and the sync layer, which answer for any component that
+  renders.
+
+  Raises Hologram.CompileError when a registered query reads an entity type declaring no allow
+  lines - default deny makes it statically dead, returning no rows when it is the query's root and
+  no embedded row when it is an include target - and when a registered query filters or orders on
+  a server-only attribute, which the client never holds and so could never evaluate locally.
+  """
+  @spec build_queries(list(module), list(module)) :: %{
+          entries: %{String.t() => %{atom => any}},
+          prop_params: %{{module, atom} => list(atom | nil)},
+          windows: %{String.t() => %{atom => any}}
+        }
+  def build_queries(component_modules, entity_types) do
+    module_queries =
+      Enum.map(component_modules, &{&1, QueryExtractor.extract_module_queries(&1, entity_types)})
+
+    Enum.each(module_queries, &validate_readable_queries!/1)
+    Enum.each(module_queries, &validate_client_evaluable_queries!/1)
+
+    terms = Enum.flat_map(module_queries, fn {_module, module_terms} -> module_terms end)
+    entries = Registry.build(terms)
+
+    windows =
+      entries
+      |> Map.new(fn {_id, entry} -> {entry.window_id, entry.window} end)
+      |> put_grants_window()
+
+    %{entries: entries, prop_params: prop_params_index(component_modules), windows: windows}
+  end
+
+  @doc """
+  Builds the registered queries PLT, whose items are the entries, prop params and windows a
+  running app reads, and returns it with the path to dump it at.
+  """
+  @spec build_queries_plt(%{atom => any}, T.opts()) :: {PLT.t(), T.file_path()}
+  def build_queries_plt(queries, opts) do
+    plt = PLT.start(items: Map.to_list(queries), supervisor: opts[:supervisor])
+    dump_path = Path.join([opts[:build_dir], Reflection.queries_plt_dump_file_name()])
+
+    {plt, dump_path}
+  end
+
+  @doc """
   Builds JavaScript code for the given Hologram page.
 
   Benchmark: https://github.com/bartblast/hologram/blob/master/benchmarks/elixir/compiler/build_page_js_6/README.md
@@ -1207,6 +1257,18 @@ defmodule Hologram.Compiler do
     |> Map.new()
   end
 
+  # The same params `prop_params/1` collects for the bundle, keyed by the pair a renderer asks
+  # about rather than by module - the bundle ships one entry per component and reads it whole,
+  # while a render looks up one prop of one component at a time.
+  defp prop_params_index(component_modules) do
+    component_modules
+    |> prop_params()
+    |> Enum.flat_map(fn {module, params} ->
+      Enum.map(params, fn {prop_name, param_names} -> {{module, prop_name}, param_names} end)
+    end)
+    |> Map.new()
+  end
+
   # Any literal is resolved, composites included, as long as every part of it is one too - a single
   # expression anywhere inside makes the whole value unknowable until it runs. Pids, ports and
   # references can't be written in a template at all (they only come from calls, which aren't
@@ -1810,6 +1872,134 @@ defmodule Hologram.Compiler do
     if beam_path != :non_existing do
       beam_path
     end
+  end
+
+  defp dead_include_message(module, dead_entity_types) do
+    "the registered query in #{inspect(module)} includes #{listing(dead_entity_types)}, which #{declare_verb(dead_entity_types)} no allow lines - default deny leaves the embed empty in every row. Add allow lines, or drop the include."
+  end
+
+  defp dead_root_message(module, dead_entity_types) do
+    "the registered query in #{inspect(module)} reads #{listing(dead_entity_types)}, which #{declare_verb(dead_entity_types)} no allow lines - default deny returns no rows to any session. Add allow lines, or drop the query."
+  end
+
+  defp declare_verb([_single_entity_type]), do: "declares"
+
+  defp declare_verb(_entity_types), do: "declare"
+
+  defp included_entity_types(term) do
+    term.include
+    |> Map.values()
+    |> Enum.flat_map(&queried_entity_types/1)
+    |> Enum.uniq()
+  end
+
+  defp listing(entity_types) do
+    Enum.map_join(entity_types, ", ", &inspect/1)
+  end
+
+  # Registered whenever the app HAS a grant store, because this is a lookup table: what decides
+  # whether any client subscribes is the BUILD, which knows whether a page can check permissions
+  # locally. An entry nothing asks for costs one map key, where a missing entry would answer
+  # :no_window to a client the build did tell to ask.
+  #
+  # An app designating no user entity type has no grant store - the grant entity joins the data
+  # model only with one - so there is no table to evaluate the window against and nothing to
+  # register.
+  defp put_grants_window(windows) do
+    if Reflection.user_entity() do
+      window = Auth.grants_window()
+
+      Map.put_new(windows, Registry.id(window), window)
+    else
+      windows
+    end
+  end
+
+  defp queried_entity_types(term) do
+    nested_types =
+      term.include
+      |> Map.values()
+      |> Enum.flat_map(&queried_entity_types/1)
+
+    Enum.uniq([term.entity | nested_types])
+  end
+
+  defp server_only_listing(references) do
+    Enum.map_join(references, ", ", fn {entity_type, names} ->
+      "#{inspect(entity_type)} #{Enum.map_join(names, ", ", &inspect/1)}"
+    end)
+  end
+
+  defp server_only_query_message(module, references) do
+    "the registered query in #{inspect(module)} filters or orders on server_only attributes (#{server_only_listing(references)}) - the client never holds those values, so it could not evaluate the reference locally. Drop the reference, or read the rows through the trusted backend API."
+  end
+
+  # Pairs each term entity with the server-only attributes its own filter and order_by name,
+  # walking include sub-terms so a reference nested under an include is reached too.
+  defp server_only_references(term) do
+    server_only_names = Entity.server_only_attribute_names(term.entity)
+    filter_names = Enum.map(term.filter, fn {name, _operator, _value} -> name end)
+    order_by_names = Enum.map(term.order_by, fn {name, _direction} -> name end)
+
+    referenced_names =
+      (filter_names ++ order_by_names)
+      |> Enum.filter(&(&1 in server_only_names))
+      |> Enum.uniq()
+      |> Enum.sort()
+
+    nested_references =
+      term.include
+      |> Map.values()
+      |> Enum.flat_map(&server_only_references/1)
+
+    if referenced_names == [] do
+      nested_references
+    else
+      [{term.entity, referenced_names} | nested_references]
+    end
+  end
+
+  # A registered query is provisioned to the client, which evaluates it locally over rows that
+  # never carry a server-only value - so it must not reference one.
+  defp validate_client_evaluable_queries!({module, module_terms}) do
+    Enum.each(module_terms, fn term ->
+      case server_only_references(term) do
+        [] ->
+          :ok
+
+        references ->
+          raise Hologram.CompileError, message: server_only_query_message(module, references)
+      end
+    end)
+  end
+
+  defp validate_readable_includes!(module, term) do
+    dead_entity_types =
+      term
+      |> included_entity_types()
+      |> Policy.dead_entity_types()
+
+    if dead_entity_types != [] do
+      raise Hologram.CompileError, message: dead_include_message(module, dead_entity_types)
+    end
+
+    :ok
+  end
+
+  # A registered query naming an entity type with no allow lines is statically dead: the policied
+  # read path composes default deny into every statement it reaches. The root and an include target
+  # fail differently, so they are checked and reported apart - and a dead root is reported alone,
+  # because a query returning no rows produces no embeds to be empty either.
+  defp validate_readable_queries!({module, module_terms}) do
+    Enum.each(module_terms, fn term ->
+      case Policy.dead_entity_types([term.entity]) do
+        [] ->
+          validate_readable_includes!(module, term)
+
+        dead_entity_types ->
+          raise Hologram.CompileError, message: dead_root_message(module, dead_entity_types)
+      end
+    end)
   end
 
   defp grants_window_id do
