+++ test/elixir/hologram/db/query_cache_test.exs
@@ -12,12 +12,6 @@ defmodule Hologram.DB.QueryCacheTest do
   alias Hologram.Query.Placeholder
   alias Hologram.Query.Registry
   alias Hologram.Test.Fixtures.Component.Module11
-  alias Hologram.Test.Fixtures.Component.Module15
-  alias Hologram.Test.Fixtures.Component.Module16
-  alias Hologram.Test.Fixtures.Component.Module17
-  alias Hologram.Test.Fixtures.Component.Module18
-  alias Hologram.Test.Fixtures.Component.Module19
-  alias Hologram.Test.Fixtures.Component.Module20
   alias Hologram.Test.Fixtures.Entity.Module2, as: Entity2
 
   use_module_stub :query_cache
@@ -98,80 +92,6 @@ defmodule Hologram.DB.QueryCacheTest do
     assert init(nil) == {:ok, nil}
 
     assert :persistent_term.get(QueryCacheStub.persistent_term_key()) == expected_data()
-  end
-
-  test "init/1 raises for a registered query whose root type declares no allow lines" do
-    stub(QueryCacheMock, :component_modules, fn -> [Module15] end)
-
-    expected_msg =
-      "the registered query in Hologram.Test.Fixtures.Component.Module15 reads " <>
-        "Hologram.Test.Fixtures.Entity.Module1, which declares no allow lines - " <>
-        "default deny returns no rows to any session. Add allow lines, or drop the query."
-
-    assert_error Hologram.CompileError, expected_msg, fn ->
-      init(nil)
-    end
-  end
-
-  test "init/1 raises for a registered query whose include target declares no allow lines" do
-    stub(QueryCacheMock, :component_modules, fn -> [Module16] end)
-
-    expected_msg =
-      "the registered query in Hologram.Test.Fixtures.Component.Module16 includes " <>
-        "Hologram.Test.Fixtures.Entity.Module1, which declares no allow lines - " <>
-        "default deny leaves the embed empty in every row. Add allow lines, or drop the include."
-
-    assert_error Hologram.CompileError, expected_msg, fn ->
-      init(nil)
-    end
-  end
-
-  test "init/1 populates a registered query that reads a type with server-only attributes without referencing them" do
-    stub(QueryCacheMock, :component_modules, fn -> [Module20] end)
-
-    assert init(nil) == {:ok, nil}
-  end
-
-  test "init/1 raises for a registered query filtering on a server-only attribute" do
-    stub(QueryCacheMock, :component_modules, fn -> [Module17] end)
-
-    expected_msg =
-      "the registered query in Hologram.Test.Fixtures.Component.Module17 filters or orders on " <>
-        "server_only attributes (Hologram.Test.Fixtures.Entity.Module15 :token) - the client " <>
-        "never holds those values, so it could not evaluate the reference locally. Drop the " <>
-        "reference, or read the rows through the trusted backend API."
-
-    assert_error Hologram.CompileError, expected_msg, fn ->
-      init(nil)
-    end
-  end
-
-  test "init/1 raises for a registered query ordering on a server-only attribute" do
-    stub(QueryCacheMock, :component_modules, fn -> [Module18] end)
-
-    expected_msg =
-      "the registered query in Hologram.Test.Fixtures.Component.Module18 filters or orders on " <>
-        "server_only attributes (Hologram.Test.Fixtures.Entity.Module15 :token) - the client " <>
-        "never holds those values, so it could not evaluate the reference locally. Drop the " <>
-        "reference, or read the rows through the trusted backend API."
-
-    assert_error Hologram.CompileError, expected_msg, fn ->
-      init(nil)
-    end
-  end
-
-  test "init/1 raises for a registered query filtering on a server-only attribute inside an include" do
-    stub(QueryCacheMock, :component_modules, fn -> [Module19] end)
-
-    expected_msg =
-      "the registered query in Hologram.Test.Fixtures.Component.Module19 filters or orders on " <>
-        "server_only attributes (Hologram.Test.Fixtures.Entity.Module15 :token) - the client " <>
-        "never holds those values, so it could not evaluate the reference locally. Drop the " <>
-        "reference, or read the rows through the trusted backend API."
-
-    assert_error Hologram.CompileError, expected_msg, fn ->
-      init(nil)
-    end
   end
 
   describe "prop_params/2" do
